+++ .gitignore
@@ -1,11 +1,32 @@
-node_modules
-dist
-.cache
 tsc-dist
+.cache
 .env
-static/*.json
-static/*.csv
+public/*.json
+public/*.csv
 .prettierrc
 settings.json
+
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
+
+node_modules
+dist
+dist-ssr
+*.local
+
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea
 .DS_Store
-.parcel-cache
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
