+++ build-data.js
@@ -4,7 +4,7 @@ const dotenv = require("dotenv");
 
 dotenv.config();
 
-require("./tsc-dist/lib/run-queries").main().catch(e => {
+require("./tsc-dist/run-queries").main().catch(e => {
   console.log(e);
   process.exit(1);
 });
