+++ index.html
@@ -1,4 +1,5 @@
-<html>
+<!doctype html>
+<html lang="en">
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
@@ -84,6 +85,6 @@
   </head>
   <body>
     <div id="app"><div class="container">Loading&hellip;</div></div>
-    <script src="./main.tsx" type="module"></script>
+    <script src="./src/main.tsx" type="module"></script>
   </body>
 </html>
