+++ package.json
@@ -1,50 +1,43 @@
 {
-  "name": "dac-viz",
+  "name": "rtc-eviction-viz",
+  "scripts": {
+    "dev": "vite",
+    "build-data": "tsc --build tsconfig.script.json && node build-data.js",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
+  },
   "dependencies": {
-    "@justfixnyc/util": "^0.1.1-alpha.0",
-    "@types/node": "^14.14.8",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "concurrently": "^5.3.0",
-    "csv-stringify": "^5.5.3",
-    "dotenv": "^12.0.3",
-    "flat": "^5.0.2",
-    "parcel": "^2.10.0",
-    "parcel-reporter-static-files-copy": "^1.5.3",
-    "pg": "^8.5.1",
-    "pg-promise": "^10.8.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-minimal-pie-chart": "^8.2.0",
-    "react-router-dom": "6",
-    "react-table": "^7.7.0",
-    "react-use-cookie": "^1.3.0",
-    "source-map-support": "^0.5.19",
-    "typescript": "^4.5.4",
+    "@justfixnyc/util": "^0.4.1",
+    "@types/node": "^20.8.10",
+    "@types/react-table": "^7.7.17",
+    "concurrently": "^8.2.2",
+    "csv-stringify": "^6.4.4",
+    "dotenv": "^16.3.1",
+    "flat": "^6.0.1",
+    "pg": "^8.11.3",
+    "pg-promise": "^11.5.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-minimal-pie-chart": "^8.4.0",
+    "react-router-dom": "^6.18.0",
+    "react-table": "^7.8.0",
+    "react-use-cookie": "^1.4.0",
+    "source-map-support": "^0.5.21",
     "vega": "^5.25.0",
     "vega-embed": "^6.23.0",
-    "vega-lite": "^4.17.0"
+    "vega-lite": "^5.16.1"
   },
-  "license": "BSD-3-Clause",
-  "staticFiles": {
-    "staticPath": "static",
-    "watcherGlob": "**"
-  },
-  "scripts": {
-    "parcel": "parcel",
-    "build-data": "tsc --build tsconfig-nodejs.json && node build-data.js",
-    "watch": "concurrently --kill-others \"parcel index.html\" \"tsc --build tsconfig-nodejs.json --watch --preserveWatchOutput\"",
-    "build": "tsc --build tsconfig-nodejs.json && parcel build index.html --public-url=./ --no-optimize",
-    "deploy": "yarn build && gh-pages -d dist"
-  },
-  "browserslist": [
-    "last 1 Chrome version"
-  ],
   "devDependencies": {
-    "@types/flat": "^5.0.1",
-    "@types/react-table": "^7.7.1",
-    "buffer": "^5.5.0||^6.0.0",
-    "gh-pages": "^3.1.0",
-    "process": "^0.11.10"
+    "@types/react": "^18.2.33",
+    "@types/react-dom": "^18.2.14",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@vitejs/plugin-react": "^4.0.3",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "typescript": "^5.0.2",
+    "vite": "^4.4.5"
   }
 }
