+++ src/constants.ts
@@ -1,0 +1,12 @@
+export const VIEW_WIDGET = "widget";
+export const VIEW_CONFIGURE_WIDGET = "config";
+
+export const QS_VIEW = "view";
+export const QS_FIELD_NAME = "fieldName";
+export const QS_HEIGHT = "height";
+
+export const EVICTION_VIZ_DEFAULT_HEIGHT = 150;
+export const ACTIVE_CASES_VIZ_DEFAULT_HEIGHT = 500;
+export const JUDGMENTS_VIZ_DEFAULT_HEIGHT = 500;
+
+export const REACT_APP_PASSWORD = process.env.REACT_APP_PASSWORD || ""
