+++ src/lib/filings-by-zip-table-outside-nyc/viz.tsx
@@ -1,3 +1,5 @@
+// @ts-nocheck
+
 import React from "react";
 import { useTable, Column, useGroupBy, useExpanded, CellProps} from "react-table";
 import { FilingsByZipOutsideNYCRow, FILINGS_BY_ZIP_OUTSIDE_NYC_TABLE } from "./data";
@@ -56,11 +58,11 @@ const Table: React.FC<Props> = ({ columns, data }) => {
                 ))}
             </thead>
             <tbody {...getTableBodyProps()}>
-                {rows.map((row, i) => {
+                {rows.map((row:any, i:number) => {
                 prepareRow(row);
                 return (
                     <tr {...row.getRowProps()}>
-                    {row.cells.map(cell => {
+                    {row.cells.map((cell:any) => {
                         return (
                             <td align="left"
                                 {...cell.getCellProps()}
