+++ src/lib/monthly-filings-table-citywide/viz.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 
 import React from "react";
 import { useTable, Column, useGroupBy, useExpanded, CellProps} from "react-table";
@@ -65,11 +66,11 @@ const Table: React.FC<Props> = ({ columns, data }) => {
                 ))}
             </thead>
             <tbody {...getTableBodyProps()}>
-                {rows.map((row, i) => {
+                {rows.map((row:any, i:number) => {
                 prepareRow(row);
                 return (
                     <tr {...row.getRowProps()}>
-                    {row.cells.map(cell => {
+                    {row.cells.map((cell:any) => {
                         return (
                             <td align="left"
                                 {...cell.getCellProps()}
