+++ src/lib/share-represented/viz.tsx
@@ -167,6 +167,7 @@ const PctReppedVizWithValues: React.FC<PctReppedVizProps & {
               },
               opacity: {
                 condition: {
+                  // @ts-ignore
                   selection: "index",
                   value: 1,
                 },
