+++ src/lib/total-judgments-statewide/data.ts
