+++ src/lib/vega-lazy.tsx
