--- tsconfig-nodejs.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",                     /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', 'es2020', or 'ESNext'. */
-    "noEmit": false,
-    "outDir": "./tsc-dist"                        /* Redirect output structure to the directory. */
-  }
-}
